PRD - C1431 - Apoio Implantação (Divergencia MR_ID MR_grmn) + 32 bits estouro
	ACTWEB 33.309.395.118
	WEBDLL 7.309.395.51

+++ branches/LFSistemas.VLI.ACTWeb.Controllers/MacroController.cs
@@ -199,7 +199,7 @@ namespace LFSistemas.VLI.ACTWeb.Controllers
         /// <param name="id">Identificados da macro</param>
         /// <param name="tipo">Tipo da macro [ E = enviadas | R = recebidas ]</param>
         /// <returns>Retorna uma macro de acordo com os parâmetros informados</returns>
-        public Macro ObterPorId(int id, string tipo)
+        public Macro ObterPorId(Int64 id, string tipo)
         {
             return dao.ObterPorId(id, tipo);
         }
